#68 - Moved action, attribute and damage condition details components
