+++ apps/spa-3ncount3r/src/components/elements/actionDetails/actionDetails.tsx
