+++ apps/spa-3ncount3r/src/components/elements/attributeDetails/attributeDetails.tsx
