+++ apps/spa-3ncount3r/src/components/elements/damageConditionType/damageConditionType.tsx
